Make different links based on the type of webserver

+++ tugboat/init.sh
@@ -55,7 +55,14 @@ EOF
 
 # Make sure previews work when tugboat_url is set to subpath
 if [ "$TUGBOAT_URL" == "http://$TUGBOAT_DOMAIN/$TUGBOAT_TAG-$TUGBOAT_TOKEN" ] || [ "$TUGBOAT_URL" == "https://$TUGBOAT_DOMAIN/$TUGBOAT_TAG-$TUGBOAT_TOKEN" ]; then
-    rm /var/www/html
-    mkdir /var/www/html
-    ln -s /var/lib/tugboat/docroot /var/www/html/$TUGBOAT_TAG-$TUGBOAT_TOKEN
+    if [[ $TUGBOAT_IMAGE == apache* ]]; then
+        rm /var/www/html
+        mkdir /var/www/html
+        ln -s /var/lib/tugboat/docroot /var/www/html/$TUGBOAT_TAG-$TUGBOAT_TOKEN
+    fi
+    if [[ $TUGBOAT_IMAGE == nginx* ]]; then
+        rm /usr/share/nginx/html
+        mkdir /usr/share/nginx/html
+        ln -s /var/lib/tugboat/docroot /usr/share/nginx/html/$TUGBOAT_TAG-$TUGBOAT_TOKEN
+    fi
 fi
